+++ src/main.rs
@@ -1,21 +1,10 @@
-use anyhow::Result;
 use clap::Parser;
-use env_logger::{Builder as LoggerBuilder, Env, DEFAULT_FILTER_ENV};
-use log::{debug, trace};
+use color_eyre::eyre::Result;
 use neverest::{cli::Cli, printer::StdoutPrinter};
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    #[cfg(not(target_os = "windows"))]
-    if let Err((_, err)) = coredump::register_panic_handler() {
-        debug!("cannot register coredump panic handler: {err}");
-        trace!("{err:?}");
-    }
-
-    LoggerBuilder::new()
-        .parse_env(Env::new().filter_or(DEFAULT_FILTER_ENV, "warn"))
-        .format_timestamp(None)
-        .init();
+    neverest::tracing::install()?;
 
     let cli = Cli::parse();
     let mut printer = StdoutPrinter::new(cli.output, cli.color);